+++ runner/internal/workspacestore/storage_fiemap_darwin.go
@@ -1,0 +1,7 @@
+package workspacestore
+
+import "os"
+
+func observeExclusiveBytes(_ *os.File) (*int64, string) {
+	return nil, "fiemap_unsupported"
+}